requirements.txt
Flask==3.0.0
python-dotenv==1.0.0
psycopg2-binary==2.9.7
pg8000==1.30.2
requests==2.32.3
aiohttp==3.9.1

# Security system
Flask-WTF==1.1.1
Flask-Limiter==3.5.0
redis==5.0.1

# Testing and development
safety==2.3.5
coverage==7.3.2
pytest==7.4.3
pytest-flask==1.3.0

# Новые зависимости для модульной архитектуры
click==8.1.7
Werkzeug==3.0.1
Jinja2==3.1.2